requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
vanna
psycopg2-binary
langchain_community
langchain_openai
langchain_core
